Simplify nested exception handler regression test
Replace the async dependency-based repro with a synchronous nested-finally case that exercises the same handler reachability bug. Track the test as an expected ILTrim limitation because ILTrim does not implement ipconstprop.



Copilot-Session: 2cee0e07-8e06-4275-b1de-508fe931bf95
