+++ src/coreclr/tools/ILTrim.Tests/ILTrimExpectedFailures.txt
@@ -454,6 +454,7 @@ UnreachableBlock.InstanceMethodSubstitutions
 UnreachableBlock.MethodArgumentPropagation
 UnreachableBlock.MethodWithParametersSubstitutions
 UnreachableBlock.MultiStageRemoval
+UnreachableBlock.NestedFinallyInFinallyHandler
 UnreachableBlock.ReplacedJumpTarget
 UnreachableBlock.ReplacedReturns
 UnreachableBlock.ResultInliningNotPossible
